Consolidate common functionality. Move pre- and post-execute tasks into aAction and aRdbmsDestinationAction
