+++ classes/ETL/Ingestor/aIngestor.php
@@ -149,33 +149,7 @@ abstract class aIngestor extends aRdbmsDestinationAction
                                   'records_examined' => $totalRecordsProcessed,
                                   'records_loaded'   => $totalRecordsProcessed
                                   ));
-    }
-
-    /* ------------------------------------------------------------------------------------------
-     * Perform any pre-execution tasks. For example, disabling table keys on MyISAM tables, or other
-     * setup tasks.
-     *
-     * NOTE: This method must check if we are in DRYRUN mode before executing any tasks.
-     *
-     * @return true on success
-     * ------------------------------------------------------------------------------------------
-     */
-
-    abstract protected function performPreExecuteTasks();
-
-    /* ------------------------------------------------------------------------------------------
-     * Perform any post-execution tasks. For example, enabling table keys on MyISAM tables, or
-     * tracking table history.
-     *
-     * NOTE: This method must check if we are in DRYRUN mode before executing any tasks.
-     *
-     * @param $numRecordsProcessed The number of records processed during this period.
-     *
-     * @return true on success
-     * ------------------------------------------------------------------------------------------
-     */
-
-    abstract protected function performPostExecuteTasks($numRecordsProcessed);
+    }  // execute()
 
     /* ------------------------------------------------------------------------------------------
      * Perform the actual work of ingestion.
